requirements.txt
-e git+ssh://git@github.com/aclark4life/django-mongodb.git@4bde24a7e208bc582eab54a0cdc672d2473f1d45#egg=django_mongodb
Django==5.1
Faker==27.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.2
Pygments==2.18.0
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
babel==2.16.0
binaryornot==0.4.4
boto3==1.34.160
botocore==1.34.160
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
confusable-homoglyphs==3.3.1
cookiecutter==2.6.0
crispy-bootstrap5==2024.2
cryptography==43.0.0
diff-match-patch==20230430
dj-database-url==2.2.0
dj-rest-auth==6.0.0
dj-stripe==2.8.4
django-after-response==0.2.2
django-allauth==64.0.0
django-appconf==1.0.6
django-ckeditor==6.7.1
django-colorful==1.3
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.3
django-cryptography-django5==2.2
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-hijack==3.6.0
django-honeypot==1.2.1
django-imagekit==5.0.0
django-import-export==4.1.1
django-ipware==7.0.1
django-js-asset==2.2.0
django-multiselectfield==0.1.13
django-ninja==1.2.2
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-recurrence==1.11.1
django-registration==3.4
django-richtextfield==1.6.2
django-sendgrid-v5==1.2.3
django-social-share==2.3.0
django-sql-explorer==5.1.1
django-storages==1.14.4
django-tables2==2.7.0
django-timezone-field==7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.21.2
enmerkar==0.7.1
gunicorn==23.0.0
html2docx==1.6.0
icalendar==5.0.13
idna==3.7
jmespath==1.0.1
lxml==5.3.0
mailchimp-marketing==3.0.80
mailchimp-transactional==1.0.56
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
phonenumbers==8.13.43
pilkit==3.0
pillow==10.4.0
pipdeptree==2.23.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydotplus==2.0.2
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-http-client==3.3.7
python-ipware==3.0.0
python-slugify==8.0.4
python-webpack-boilerplate==1.0.3
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
stripe==4.2.0
tablib==3.5.0
text-unidecode==1.3
texttable==1.7.0
tinycss2==1.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
wheel==0.44.0
